Add a safe wrapper for mts_array_t in Rust
This should reduce the number of time a Rust user has to think
about the C API
