+++ rust/metatensor/src/data/array_ref.rs
@@ -1,8 +1,16 @@
-use crate::c_api::mts_array_t;
-use crate::data::origin::get_data_origin;
+use std::ptr::NonNull;
+
 use ndarray::{ArcArray, IxDyn};
 
+use dlpk::sys::DLDevice;
+
+use crate::c_api::{mts_array_t, mts_data_origin_t, mts_data_movement_t};
+use crate::Error;
+
 use super::Array;
+
+use super::external::check_status_external;
+use super::origin::get_data_origin;
 
 /// Reference to a data array in metatensor-core
 ///
@@ -24,7 +32,13 @@ impl<'a> ArrayRef<'a> {
     /// owner of this `mts_array_t`.
     pub unsafe fn from_raw(array: mts_array_t) -> ArrayRef<'a> {
         ArrayRef {
-            array,
+            array: mts_array_t {
+                // remove the destructor if any, since we only have a refernce
+                // to the array, and it should not be dropped when passed back
+                // to C through `as_raw()`.
+                destroy: None,
+                ..array
+            },
             marker: std::marker::PhantomData,
         }
     }
@@ -35,7 +49,7 @@ impl<'a> ArrayRef<'a> {
     /// the [`Array`] trait.
     #[inline]
     pub fn as_any(&self) -> &dyn std::any::Any {
-        let origin = self.array.origin().unwrap_or(0);
+        let origin = self.origin().unwrap_or(0);
         assert_eq!(
             origin, *super::array::RUST_DATA_ORIGIN,
             "this array was not created as a rust Array (origin is '{}')",
@@ -55,7 +69,7 @@ impl<'a> ArrayRef<'a> {
     /// the [`Array`] trait.
     #[inline]
     pub fn to_any(self) -> &'a dyn std::any::Any {
-        let origin = self.array.origin().unwrap_or(0);
+        let origin = self.origin().unwrap_or(0);
         assert_eq!(
             origin, *super::array::RUST_DATA_ORIGIN,
             "this array was not created as a rust Array (origin is '{}')",
@@ -81,13 +95,150 @@ impl<'a> ArrayRef<'a> {
     /// This function will panic if the data in this `mts_array_t` is not a
     /// `ndarray::ArcArray`.
     #[inline]
-    pub fn to_array(self) -> &'a ArcArray<f64, IxDyn> {
+    pub fn to_ndarray(self) -> &'a ArcArray<f64, IxDyn> {
         self.to_any().downcast_ref().expect("this is not a ndarray::ArcArray")
     }
 
     /// Get the raw underlying `mts_array_t`
     pub fn as_raw(&self) -> &mts_array_t {
         &self.array
+    }
+
+    /// Get the origin of this array.
+    ///
+    /// This corresponds to `mts_array_t.origin`, but with a more convenient API.
+    pub fn origin(&self) -> Result<mts_data_origin_t, Error> {
+        let function = self.array.origin.expect("mts_array_t.origin function is NULL");
+
+        let mut origin = 0;
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut origin),
+                "mts_array_t.origin",
+            )?;
+        }
+
+        return Ok(origin);
+    }
+
+    /// Get the device of this array.
+    ///
+    /// This corresponds to `mts_array_t.device`, but with a more convenient API.
+    pub fn device(&self) -> Result<DLDevice, Error> {
+        let function = self.array.device.expect("mts_array_t.device function is NULL");
+
+        let mut device = DLDevice::cpu();
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut device),
+                "mts_array_t.device",
+            )?;
+        }
+
+        return Ok(device);
+    }
+
+    /// Get the dtype of this array.
+    ///
+    /// This corresponds to `mts_array_t.dtype`, but with a more convenient API.
+    pub fn dtype(&self) -> Result<dlpk::sys::DLDataType, Error> {
+        let function = self.array.dtype.expect("mts_array_t.dtype function is NULL");
+
+        let mut dtype = dlpk::sys::DLDataType { code: dlpk::sys::DLDataTypeCode::kDLFloat, bits: 0, lanes: 0 };
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut dtype),
+                "mts_array_t.dtype",
+            )?;
+        }
+
+        return Ok(dtype);
+    }
+
+    /// Get a [`dlpk::DLPackTensor`] from this array, if supported by the underlying data.
+    ///
+    /// This corresponds to `mts_array_t.as_dlpack`, but with a more convenient API.
+    pub fn as_dlpack(
+        &self,
+        device: DLDevice,
+        stream: Option<i64>,
+        max_version: dlpk::sys::DLPackVersion,
+    ) -> Result<dlpk::DLPackTensor, Error> {
+        let function = self.array.as_dlpack.expect("mts_array_t.as_dlpack function is NULL");
+
+        let mut tensor = std::ptr::null_mut();
+        let stream_c = stream.as_ref().map_or(std::ptr::null(), |s| s as *const i64);
+
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut tensor, device, stream_c, max_version),
+                "mts_array_t.as_dlpack",
+            )?;
+        }
+
+        let tensor = NonNull::new(tensor).expect("got a NULL DLManagedTensorVersioned from `as_dlpack`");
+        let tensor = unsafe {
+            dlpk::DLPackTensor::from_ptr(tensor)
+        };
+
+        return Ok(tensor);
+    }
+
+    /// Get the shape of this array.
+    ///
+    /// This corresponds to `mts_array_t.shape`, but with a more convenient API.
+    pub fn shape(&self) -> Result<&[usize], Error> {
+        let function = self.array.shape.expect("mts_array_t.shape function is NULL");
+
+        let mut shape = std::ptr::null();
+        let mut shape_count: usize = 0;
+
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut shape, &mut shape_count),
+                "mts_array_t.shape"
+            )?;
+        }
+
+        assert!(shape_count > 0);
+        let shape = unsafe {
+            std::slice::from_raw_parts(shape, shape_count)
+        };
+
+        return Ok(shape);
+    }
+
+    /// Create a new array with the same data as this array, but with a different shape.
+    ///
+    /// This corresponds to `mts_array_t.create`, but with a more convenient API.
+    pub fn create(&self, shape: &[usize]) -> Result<mts_array_t, Error> {
+        let function = self.array.create.expect("mts_array_t.create function is NULL");
+
+        let mut new_array = mts_array_t::null();
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, shape.as_ptr(), shape.len(), &mut new_array),
+                "mts_array_t.create",
+            )?;
+        }
+
+        return Ok(new_array);
+    }
+
+    /// Copy the data in this array, if supported by the underlying data.
+    ///
+    /// This corresponds to `mts_array_t.copy`, but with a more convenient API.
+    pub fn copy(&self) -> Result<mts_array_t, Error> {
+        let function = self.array.copy.expect("mts_array_t.copy function is NULL");
+        let mut new_array = mts_array_t::null();
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut new_array),
+                "mts_array_t.copy",
+            )?;
+        }
+
+        return Ok(new_array);
     }
 }
 
@@ -112,9 +263,15 @@ impl<'a> ArrayRefMut<'a> {
     /// `mts_array_t`; and that the owner is mutably borrowed by this
     /// `ArrayRefMut`.
     #[inline]
-    pub unsafe fn new(array: mts_array_t) -> ArrayRefMut<'a> {
+    pub unsafe fn from_raw(array: mts_array_t) -> ArrayRefMut<'a> {
         ArrayRefMut {
-            array,
+            array: mts_array_t {
+                // remove the destructor if any, since we only have a refernce
+                // to the array, and it should not be dropped when passed back
+                // to C through `as_raw()`.
+                destroy: None,
+                ..array
+            },
             marker: std::marker::PhantomData,
         }
     }
@@ -125,7 +282,7 @@ impl<'a> ArrayRefMut<'a> {
     /// the [`Array`] trait.
     #[inline]
     pub fn as_any(&self) -> &dyn std::any::Any {
-        let origin = self.array.origin().unwrap_or(0);
+        let origin = self.origin().unwrap_or(0);
         assert_eq!(
             origin, *super::array::RUST_DATA_ORIGIN,
             "this array was not created as a rust Array (origin is '{}')",
@@ -145,7 +302,7 @@ impl<'a> ArrayRefMut<'a> {
     /// the [`Array`] trait.
     #[inline]
     pub fn to_any(&self) -> &'a dyn std::any::Any {
-        let origin = self.array.origin().unwrap_or(0);
+        let origin = self.origin().unwrap_or(0);
         assert_eq!(
             origin, *super::array::RUST_DATA_ORIGIN,
             "this array was not created as a rust Array (origin is '{}')",
@@ -164,7 +321,7 @@ impl<'a> ArrayRefMut<'a> {
     /// the [`Array`] trait.
     #[inline]
     pub fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
-        let origin = self.array.origin().unwrap_or(0);
+        let origin = self.origin().unwrap_or(0);
         assert_eq!(
             origin, *super::array::RUST_DATA_ORIGIN,
             "this array was not created as a rust Array (origin is '{}')",
@@ -184,7 +341,7 @@ impl<'a> ArrayRefMut<'a> {
     /// the [`Array`] trait.
     #[inline]
     pub fn to_any_mut(self) -> &'a mut dyn std::any::Any {
-        let origin = self.array.origin().unwrap_or(0);
+        let origin = self.origin().unwrap_or(0);
         assert_eq!(
             origin, *super::array::RUST_DATA_ORIGIN,
             "this array was not created as a rust Array (origin is '{}')",
@@ -210,7 +367,7 @@ impl<'a> ArrayRefMut<'a> {
     /// This function will panic if the data in this `mts_array_t` is not a
     /// `ndarray::ArcArray`.
     #[inline]
-    pub fn to_array(&self) -> &ArcArray<f64, IxDyn> {
+    pub fn to_ndarray(&self) -> &ArcArray<f64, IxDyn> {
         self.to_any().downcast_ref().expect("this is not a ndarray::ArcArray")
     }
 
@@ -218,7 +375,7 @@ impl<'a> ArrayRefMut<'a> {
     /// `ndarray::ArcArray`. This function will panic if the data in this
     /// `mts_array_t` is not a `ndarray::ArcArray`.
     #[inline]
-    pub fn as_array_mut(&mut self) -> &mut ArcArray<f64, IxDyn> {
+    pub fn as_ndarray_mut(&mut self) -> &mut ArcArray<f64, IxDyn> {
         self.as_any_mut().downcast_mut().expect("this is not a ndarray::ArcArray")
     }
 
@@ -228,7 +385,7 @@ impl<'a> ArrayRefMut<'a> {
     /// This function will panic if the data in this `mts_array_t` is not a
     /// `ndarray::ArcArray`.
     #[inline]
-    pub fn to_array_mut(self) -> &'a mut ArcArray<f64, IxDyn> {
+    pub fn to_ndarray_mut(self) -> &'a mut ArcArray<f64, IxDyn> {
         self.to_any_mut().downcast_mut().expect("this is not a ndarray::ArcArray")
     }
 
@@ -240,5 +397,195 @@ impl<'a> ArrayRefMut<'a> {
     /// Get a mutable reference to the raw underlying `mts_array_t`
     pub fn as_raw_mut(&mut self) -> &mut mts_array_t {
         &mut self.array
+    }
+
+    /// Get the origin of this array.
+    ///
+    /// This corresponds to `mts_array_t.origin`, but with a more convenient API.
+    pub fn origin(&self) -> Result<mts_data_origin_t, Error> {
+        let function = self.array.origin.expect("mts_array_t.origin function is NULL");
+
+        let mut origin = 0;
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut origin),
+                "mts_array_t.origin",
+            )?;
+        }
+
+        return Ok(origin);
+    }
+
+    /// Get the device of this array.
+    ///
+    /// This corresponds to `mts_array_t.device`, but with a more convenient API.
+    pub fn device(&self) -> Result<DLDevice, Error> {
+        let function = self.array.device.expect("mts_array_t.device function is NULL");
+
+        let mut device = DLDevice::cpu();
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut device),
+                "mts_array_t.device",
+            )?;
+        }
+
+        return Ok(device);
+    }
+
+    /// Get the dtype of this array.
+    ///
+    /// This corresponds to `mts_array_t.dtype`, but with a more convenient API.
+    pub fn dtype(&self) -> Result<dlpk::sys::DLDataType, Error> {
+        let function = self.array.dtype.expect("mts_array_t.dtype function is NULL");
+
+        let mut dtype = dlpk::sys::DLDataType { code: dlpk::sys::DLDataTypeCode::kDLFloat, bits: 0, lanes: 0 };
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut dtype),
+                "mts_array_t.dtype",
+            )?;
+        }
+
+        return Ok(dtype);
+    }
+
+    /// Get a [`dlpk::DLPackTensor`] from this array, if supported by the underlying data.
+    ///
+    /// This corresponds to `mts_array_t.as_dlpack`, but with a more convenient API.
+    pub fn as_dlpack(
+        &self,
+        device: DLDevice,
+        stream: Option<i64>,
+        max_version: dlpk::sys::DLPackVersion,
+    ) -> Result<dlpk::DLPackTensor, Error> {
+        let function = self.array.as_dlpack.expect("mts_array_t.as_dlpack function is NULL");
+
+        let mut tensor = std::ptr::null_mut();
+        let stream_c = stream.as_ref().map_or(std::ptr::null(), |s| s as *const i64);
+
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut tensor, device, stream_c, max_version),
+                "mts_array_t.as_dlpack",
+            )?;
+        }
+
+        let tensor = NonNull::new(tensor).expect("got a NULL DLManagedTensorVersioned from `as_dlpack`");
+        let tensor = unsafe {
+            dlpk::DLPackTensor::from_ptr(tensor)
+        };
+
+        return Ok(tensor);
+    }
+
+    /// Get the shape of this array.
+    ///
+    /// This corresponds to `mts_array_t.shape`, but with a more convenient API.
+    pub fn shape(&self) -> Result<&[usize], Error> {
+        let function = self.array.shape.expect("mts_array_t.shape function is NULL");
+
+        let mut shape = std::ptr::null();
+        let mut shape_count: usize = 0;
+
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut shape, &mut shape_count),
+                "mts_array_t.shape"
+            )?;
+        }
+
+        assert!(shape_count > 0);
+        let shape = unsafe {
+            std::slice::from_raw_parts(shape, shape_count)
+        };
+
+        return Ok(shape);
+    }
+
+    /// Reshape the data in this array, if supported by the underlying data.
+    ///
+    /// This corresponds to `mts_array_t.reshape`, but with a more convenient API.
+    pub fn reshape(&mut self, shape: &[usize]) -> Result<(), Error> {
+        let function = self.array.reshape.expect("mts_array_t.reshape function is NULL");
+
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, shape.as_ptr(), shape.len()),
+                "mts_array_t.reshape",
+            )?;
+        }
+
+        return Ok(());
+    }
+
+    /// Swap two axes of the data in this array, if supported by the underlying data.
+    ///
+    /// This corresponds to `mts_array_t.swap_axes`, but with a more convenient API.
+    pub fn swap_axes(&mut self, axis_1: usize, axis_2: usize) -> Result<(), Error> {
+        let function = self.array.swap_axes.expect("mts_array_t.swap_axes function is NULL");
+
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, axis_1, axis_2),
+                "mts_array_t.swap_axes",
+            )?;
+        }
+
+        return Ok(());
+    }
+
+    /// Create a new array with the same data as this array, but with a different shape.
+    ///
+    /// This corresponds to `mts_array_t.create`, but with a more convenient API.
+    pub fn create(&self, shape: &[usize]) -> Result<mts_array_t, Error> {
+        let function = self.array.create.expect("mts_array_t.create function is NULL");
+
+        let mut new_array = mts_array_t::null();
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, shape.as_ptr(), shape.len(), &mut new_array),
+                "mts_array_t.create",
+            )?;
+        }
+
+        return Ok(new_array);
+    }
+
+    /// Copy the data in this array, if supported by the underlying data.
+    ///
+    /// This corresponds to `mts_array_t.copy`, but with a more convenient API.
+    pub fn copy(&self) -> Result<mts_array_t, Error> {
+        let function = self.array.copy.expect("mts_array_t.copy function is NULL");
+        let mut new_array = mts_array_t::null();
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, &mut new_array),
+                "mts_array_t.copy",
+            )?;
+        }
+
+        return Ok(new_array);
+    }
+
+    /// Move the data in this array to another array, if supported by the underlying data.
+    ///
+    /// This corresponds to `mts_array_t.move_data`, but with a more convenient API.
+    pub fn move_data<'input>(
+        &mut self,
+        input: impl Into<ArrayRef<'input>>,
+        moves: &[mts_data_movement_t],
+    ) -> Result<(), Error> {
+        let function = self.array.move_data.expect("mts_array_t.move_data function is NULL");
+
+        let input = input.into();
+        unsafe {
+            check_status_external(
+                function(self.array.ptr, input.as_raw().ptr, moves.as_ptr(), moves.len()),
+                "mts_array_t.move_data",
+            )?;
+        }
+
+        return Ok(());
     }
 }
